# dist/openapi.yaml
openapi: 3.0.0
info:
  title: recipe - web
  version: 1.0.0
  description: API for user and recipe management in a recipe web app
servers:
  - url: http://localhost:3000
    description: SwaggerHub API Auto Mocking
paths:
  /register:
    post:
      tags:
        - user
      summary: Register a new user
      description: >
        Registers a new user in the system. The username must be unique. 
        If a user tries to register with an existing username, the server will return 409 Conflict.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUser'
      responses:
        '201':
          description: User registered successfully
        '400':
          description: Invalid input (e.g. password mismatch or username too short)
        '409':
          description: Username already exists
  /login:
    post:
      tags:
        - user
      summary: Login user
      description: Authenticates a user using username and password. On success, the user is considered logged in.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
        '401':
          description: Invalid username or password
...
  /recipes/{recipeId}:
    get:
      tags:
        - recipe
      summary: Get full recipe details
      description: >
        Returns full details of a specific recipe by its ID, including image, title, duration,
        popularity (likes), dietary flags, ingredients list, instructions, and servings.
      parameters:
        - name: recipeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Full recipe details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FullRecipe'
        '404':
          description: Recipe not found
...
  /recipes:
    post:
      tags:
        - recipe
      summary: Create a new recipe
      description: >
        Allows an authenticated user to submit a new recipe. 
        The recipe will be stored in the database and shown in their personal area under "My Recipes".
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewRecipe'
      responses:
        '201':
          description: Recipe created successfully
        '400':
          description: Invalid input data
        '401':
          description: Unauthorized – user must be logged in
...
components:
  schemas:
    RegisterUser:
      type: object
      required:
        - username
        - first_name
        - last_name
        - country
        - password
        - confirm_password
        - email
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 8
          pattern: "^[A-Za-z]+$"
          example: hadaroni
        first_name:
          type: string
          pattern: "^[A-Za-z]+$"
          example: hadaroni
        last_name:
          type: string
          pattern: "^[A-Za-z]+$"
          example: hadaroni
        country:
          type: string
          example: Israel
        password:
          type: string
          minLength: 5
          maxLength: 10
          example: Pa$$w0rd
          description: Must contain at least one number and one special character
        confirm_password:
          type: string
          example: Pa$$w0rd
        email:
          type: string
          format: email
          example: haadaroni@example.com
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: hadaroni
        password:
          type: string
          example: Pa$$w0rd
...
    FullRecipe:
      type: object
      required:
        - id
        - title
        - image
        - duration
        - likes
        - isVegan
        - isVegetarian
        - isGlutenFree
        - ingredients
        - instructions
        - servings
      properties:
        id:
          type: string
          example: abc123
        title:
          type: string
          example: Vegan Stir Fry
        image:
          type: string
          format: uri
          example: https://example.com/images/stirfry.jpg
        duration:
          type: string
          example: "30 minutes"
        likes:
          type: integer
          example: 1000
        isVegan:
          type: boolean
          example: true
        isVegetarian:
          type: boolean
          example: true
        isGlutenFree:
          type: boolean
          example: false
        viewed:
          type: boolean
          example: true
        isFavorite:
          type: boolean
          example: false
        ingredients:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                example: Broccoli
              amount:
                type: string
                example: "200g"
        instructions:
          type: string
          example: "1. Cut vegetables\n2. Stir fry with sauce for 10 minutes..."
        servings:
          type: integer
          example: 4
    NewRecipe:
      type: object
      required:
        - title
        - image
        - duration
        - ingredients
        - instructions
        - servings
        - isVegan
        - isVegetarian
        - isGlutenFree
      properties:
        title:
          type: string
          example: Green Thai Curry
        image:
          type: string
          format: uri
          example: https://example.com/images/curry.jpg
        duration:
          type: integer
          example: 30
        popularity:
          type: integer
          example: 0
          description: Number of likes, initialized to 0
        isVegan:
          type: boolean
          example: true
        isVegetarian:
          type: boolean
          example: true
        isGlutenFree:
          type: boolean
          example: false
        ingredients:
          type: array
          items:
            type: object
            required:
              - name
              - quantity
            properties:
              name:
                type: string
                example: Coconut Milk
              quantity:
                type: string
                example: 1 can
        instructions:
          type: string
          example: >
            1. Heat oil in pan. 
            2. Add curry paste and fry. 
            3. Add coconut milk and vegetables.
        servings:
          type: integer
          example: 4
...